Prisma/Code/Src/Python/prisma-ai-extractors/src/extractors/paddle_extractor.py
"""
PaddleOCR extractor for document processing.
"""

from typing import Dict, Any, Optional
from PIL import Image

from .base_extractor import BaseExtractor


class PaddleOCRExtractor(BaseExtractor):
    """PaddleOCR based document extractor."""
    
    def __init__(self, config: Optional[Dict[str, Any]] = None):
        """
        Initialize PaddleOCR extractor.
        
        Args:
            config: Configuration dictionary
        """
        super().__init__(config)
    
    def _initialize(self):
        """Initialize PaddleOCR."""
        # Stub implementation
        print(f"[INFO] Initializing {self.name} (stub implementation)")
        self.model = None  # Placeholder
    
    def _extract_impl(self, image: Image.Image) -> Dict[str, Any]:
        """
        Extract information using PaddleOCR.
        
        Args:
            image: PIL Image object
            
        Returns:
            Dictionary with extracted data
        """
        # Stub implementation
        return {
            "fecha": "2024-01-01",
            "autoridadEmisora": "STUB",
            "expediente": "PADDLE-STUB",
            "note": "This is a stub implementation"
        }